+++ Network/API/Telegram/Bot/Object/Update/Message.hs
@@ -19,7 +19,7 @@ import "text" Data.Text (Text)
 import Network.API.Telegram.Bot.Object.Update.Message.Content (Content)
 import Network.API.Telegram.Bot.Object.Update.Message.Origin (Origin (Private, Group, Supergroup, Channel))
 import Network.API.Telegram.Bot.Property.Accessible (Accessible (access))
-import Network.API.Telegram.Bot.Property.Identifiable (Identifiable (Identificator, identificator))
+import Network.API.Telegram.Bot.Property.Identifiable (Identifiable (Identificator, ident))
 import Network.API.Telegram.Bot.Property.Persistable (Persistable (Payload, payload, endpoint)
 	, Capacity (Send, Edit, Purge), Inform (Notify, Silently), Way (Directly, Forwarding, Replying))
 
@@ -71,9 +71,9 @@ instance FromJSON Message where
 
 instance Identifiable Message where
 	type instance Identificator Message = Int
-	identificator (Direct i _ _) = i
-	identificator (Forward i _ _) = i
-	identificator (Reply i _ _ _) = i
+	ident (Direct i _ _) = i
+	ident (Forward i _ _) = i
+	ident (Reply i _ _ _) = i
 
 instance Persistable ('Send 'Notify 'Directly) Message where
 	type instance Payload ('Send 'Notify 'Directly) Message