+++ Network/API/Telegram/Bot/Property/Accessible.hs
@@ -3,4 +3,5 @@ module Network.API.Telegram.Bot.Property.Accessible (Accessible (..)) where
 import "lens" Control.Lens (Lens')
 
 class Accessible target source where
+	{-# MINIMAL access #-}
 	access :: Lens' source target